Refactor distribution/datatypes into a providers/datatypes
